Replaced seed type with UInt32.

+++ include/Vortex/Math/Random.h
@@ -6,8 +6,8 @@
 namespace Vortex::Math {
 	struct Random {
 	public:
-		explicit Random(UInt64 seed
-		= static_cast<long unsigned int>(std::chrono::high_resolution_clock::now().time_since_epoch().count()))
+		explicit Random(UInt32 seed
+		= static_cast<UInt32>(std::chrono::high_resolution_clock::now().time_since_epoch().count()))
 			: m_Seed{seed}, m_RandomDevice(m_Seed) {}
 		virtual ~Random() = default;
 
@@ -34,14 +34,14 @@ namespace Vortex::Math {
 				return distribution(m_RandomDevice);
 			}
 		}
-		inline void SetSeed(UInt64 seed) {
+		inline void SetSeed(UInt32 seed) {
 			m_RandomDevice.seed(seed);
 			m_Seed = seed;
 		}
-		inline UInt64 GetSeed() const { return m_Seed; }
+		inline UInt32 GetSeed() const { return m_Seed; }
 
 	private:
-		UInt64 m_Seed;
+		UInt32 m_Seed;
 		std::mt19937 m_RandomDevice;
 	};
 }
